requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
